asyncpg, more like asyncpog

+++ db/migrations/20231210172725_item.sql
@@ -2,7 +2,7 @@
 CREATE TABLE item (
     id INTEGER PRIMARY KEY,
     name VARCHAR(64) NOT NULL,
-    price INTEGER
+    price NUMERIC(10,2)
 );
 
 -- migrate:down
